Rename HTML properties on Header/Footer to better convey meaning
