--- PugPdf.Core/TextUtil.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Text;
-
-namespace PugPdf.Core
-{
-    public static class TextUtil
-    {
-        public static string SpecialCharsEncode(string text)
-        {
-            var charArray = text.ToCharArray();
-            var stringBuilder = new StringBuilder();
-
-            foreach (var ch in charArray)
-            {
-                var charInt = Convert.ToInt32(ch);
-
-                if (charInt > sbyte.MaxValue)
-                    stringBuilder.AppendFormat("&#{0};", charInt);
-                else
-                    stringBuilder.Append(ch);
-            }
-
-            return stringBuilder.ToString();
-        }
-    }
-}
